mvp-project/backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["adilet <adilet.issayev@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
sqlalchemy = "^2.0.36"
alembic = "^1.14.0"
psycopg2 = "^2.9.10"
motor = "^3.6.0"
pydantic = "^2.10.2"
fastapi = "^0.115.5"
uvicorn = "^0.32.1"
bcrypt = "^4.2.1"
python-jose = "^3.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
